KOGITO-5664 - Rename KafkaClient package to avoid Sonar conflict
