+++ process-kafka-multi-springboot/src/test/java/org/acme/travel/MultiMessagingIT.java
@@ -26,7 +26,7 @@ import java.util.stream.IntStream;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
-import org.kie.kogito.test.kafka.KafkaTestClient;
+import org.kie.kogito.test.springboot.kafka.KafkaTestClient;
 import org.kie.kogito.testcontainers.springboot.KafkaSpringBootTestResource;
 import org.kie.kogito.tests.KogitoKafkaMultiSpringbootApplication;
 import org.slf4j.Logger;
