+++ process-kafka-quickstart-springboot/src/test/java/org/acme/travel/MessagingIT.java
@@ -25,7 +25,7 @@ import java.util.stream.IntStream;
 
 import org.junit.After;
 import org.junit.jupiter.api.Test;
-import org.kie.kogito.test.kafka.KafkaTestClient;
+import org.kie.kogito.test.springboot.kafka.KafkaTestClient;
 import org.kie.kogito.testcontainers.springboot.KafkaSpringBootTestResource;
 import org.kie.kogito.tests.KogitoKafkaQuickstartSpringbootApplication;
 import org.slf4j.Logger;
